fix: buttonColor not working with old RN architecture
