+++ android/src/main/java/com/henninghall/date_picker/DatePickerModuleImpl.java
@@ -4,6 +4,7 @@ package com.henninghall.date_picker;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.graphics.Color;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
@@ -53,9 +54,10 @@ public class DatePickerModuleImpl {
         String title = props.getString("title");
         String confirmText = props.getString("confirmText");
         final String cancelText = props.getString("cancelText");
+        String buttonColor = props.getString("buttonColor");
         final View pickerWithMargin = withTopMargin(picker);
 
-        return new AlertDialog.Builder(DatePickerPackage.context.getCurrentActivity(), getTheme(props))
+        AlertDialog dialog = new AlertDialog.Builder(DatePickerPackage.context.getCurrentActivity(), getTheme(props))
                 .setTitle(title)
                 .setCancelable(true)
                 .setView(pickerWithMargin)
@@ -78,6 +80,20 @@ public class DatePickerModuleImpl {
                     }
                 })
                 .create();
+
+        dialog.setOnShowListener(new DialogInterface.OnShowListener() {
+            @Override
+            public void onShow(DialogInterface dialoga) {
+                if(buttonColor != null){
+                    int color = Color.parseColor(buttonColor);
+                    dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(color);
+                    dialog.getButton(AlertDialog.BUTTON_NEGATIVE).setTextColor(color);
+                }
+
+            }
+        });
+
+        return dialog;
     }
 
     private int getTheme(ReadableMap props) {
