+++ src/fabric/NativeRNDatePicker.ts
@@ -4,6 +4,7 @@ import { Double, UnsafeObject } from 'react-native/Libraries/Types/CodegenTypes'
 
 export interface Spec extends TurboModule {
   readonly getConstants: () => {}
+  closePicker(): void
   openPicker(props: UnsafeObject): void
   removeListeners(type: Double): void
   addListener(eventName: string): void
